Rewrite RouteAnalyser
From now on, we perform route analysis using GPX track points and
`@turf/boolean-point-in-polygon`, instead of relying on
`@turf/line-split`, which in turn relied heavily on the GeoJSON spec.
This approach gives us a very convenient way of extrating the offroute
start/end points, as well as an easy access to track point metadata.